Move the create package function in the update client
